pyproject.toml
[project]
name = "gb2seq"
version = "0.3.3"
description = "Python library and scripts for working with unannotated sequences based on an annotated reference genome provided by a GenBank file."

readme = "README.md"
authors = [
    { name = "Terry C. Jones", email = "terence.jones@charite.de" },
    { name = "Barbara Mühlemann" },
    { name = "Udo Gieraths" },
    { name = "Nikolai Zaki" },
]
requires-python = ">=3.10"
dependencies = [
    "dark-matter>=5.0.43",
    "rich>=13.9.4",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://github.com/VirologyCharite/gb2seq"
"Bug Tracker" = "https://github.com/VirologyCharite/gb2seq/issues"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
data = { scripts = "bin" }

[dependency-groups]
dev = [
    "nox>=2025.5.1",
    "pytest>=8.4.1",
]
